test: Consolidating test files.
